fix(aero-machine): align AeroGPU advertised features with implementation
- BAR0 FEATURES_LO/HI now comes from a supported_features() mask (SCANOUT/VBLANK/FENCE_PAGE only)

- Added aero-machine/tests/aerogpu_features.rs to assert feature bits via MMIO

- Fixed AeroGPU PC platform wiring: ensure 00:07.0 PCI function is present and register legacy VGA ports as exact handlers to avoid IO range overlap

- Restored build: add missing MachineError AeroGPU validation + dedupe Machine::vbe_lfb_base

- Fix aero-platform build/tests: restore PlatformInterrupts::lapics_iter helper (MSI) and repair mis-braced IOAPIC test

- Fix firmware build: pass cdrom_present into drive_present for INT 13h verify paths

- Updated vga_vbe_lfb_pci test and Cargo.lock

+++ crates/aero-machine/tests/vga_vbe_lfb_pci.rs
@@ -25,7 +25,7 @@ fn vga_vbe_lfb_is_reachable_via_pci_mmio_router() {
     // changes (e.g. standalone VGA stub vs AeroGPU BAR1-backed legacy VBE).
     let base = m.vbe_lfb_base();
     // Write a red pixel at (0,0) in packed 32bpp BGRX via *machine memory*.
-    m.write_physical_u32(u64::from(base), 0x00FF_0000);
+    m.write_physical_u32(base, 0x00FF_0000);
 
     m.display_present();
     assert_eq!(m.display_resolution(), (64, 64));
